refactor: connection db
