+++ packages/server/src/routes/index.ts
@@ -1,5 +1,6 @@
 import express from 'express'
 import User from './user'
+
 const Router = express.Router()
 
 Router.use('/users', User)
