+++ packages/server/src/routes/user.ts
@@ -1,5 +1,5 @@
 import express, { Router } from 'express'
-import UserController from '../controller/User'
+import UserController from '../controllers/User'
 const User: Router = express.Router()
 
 User.post('/create', UserController.create)
